### 2021-01-04 #2932 @arfon
:wave: @VivianePons - happy new year! Would you be willing to edit this submission for JOSS? It's not clear if this submission needs a scope check by the editorial team first so please let me know if you think we should review for scope.
