### 2021-01-04 #2932 @VivianePons
Hi @dansarie thank you very much for submitting your paper.

Could you look at our [potential reviewer list](https://bit.ly/joss-reviewers) and maybe suggest reviewers (without tagging them with an @).? You can also suggest names outside of the list, just be aware of our [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy)
