### 2021-01-04 #2932 @arfon
> @arfon I read the paper and it looks quite convincing for being on scope. What made you hesitate?

Oh nothing really. Just that ~2000 LOC (in C) might be considered on the smaller-side of things for us. Happy to proceed if you think this is 👍
