### 2021-01-07 #2932 @gradvohl
# Review itens for JOSS paper #2932
**Paper title**: sboxgates: A program for finding low gate count implementations of S-boxes

## Review criteria
* A list of the authors of the software: :heavy_check_mark:
* Author affiliations: :heavy_check_mark:
* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience: :heavy_check_mark:
* A clear statement of need that illustrates the purpose of the software: :heavy_check_mark:
* Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it: :heavy_check_mark:
* A list of key references including a link to the software archive:  :heavy_check_mark:

## Review items
* Software license: :thumbsup: (good - GPL v3)

* Documentation:
    * A statement of need: :thumbsup: (good)
    * Installation instructions: :wave: (sort of -- Please see the following Compilation section)
    * Example usage: :wave: (sort of)
    * API documentation: :thumbsup: (good -- Please see the following Documentation section)
    * Community guidelines: :thumbsup: (good)
    * Functionality: :wave: (sort of -- Please see the following Compilation section).
    * Tests: :thumbsdown: (bad)

* Software paper
    * Authors: :heavy_check_mark:
    * A statement of need: :heavy_check_mark:
    * References: :heavy_check_mark:

## Reviewer comments
The following comments refer to items I considered  that could be improved a little more, although they currently met the requirements for JOSS submission.

### Documentation in source code
The author comments on some functions in the .h files, while other comments are in .c files. Also, most of the comments describe the functions at a high level only. It would be interesting to say what the parameters are and what they are needed for.

**Suggestion**: Uniform where you will write the comments and, if possible, use a standard for documentation of functions, parameters, and returns (such as [Doxygen](https://www.doxygen.nl)).

### Compilation
- When I tried to build the software with `cmake` it required the CMake 3.10 or higher.

- **Suggestion**: Inform the user what are the required versions (especially CMake).

- When I tried to `cmake ..`, it raised the following error:
`-- Could NOT find MPI_C (missing: MPI_mpich_LIBRARY MPI_opa_LIBRARY MPI_mpl_LIBRARY MPI_C_WORKS)`

- **Suggestion**: Inform the user what are the main environment variables must be set. Besides, I suspect that `mpich-devel` is also a dependency.

### Questions:
When running with mpirun, may the user specify the number of processes and `hostfile` with processes location (for instance, `mpirun -np 4 --hostfile file ...`? If so, the author should clarify this in the instructions for running the software.

 ### Other suggestions:
1. Please consider to add the author's name and contact info in the `README.md`.
2. In `paper.md`, the author states that CUDA is a "programming language for GPUs". However, CUDA is an Application Programming Interface (API) or a library, if you prefer.

## Conclusion
Analyzing all the aspects mentioned earlier, I rank the paper as **minor revisions** needed. I would like to ask the author to provide the corrections, especially in the compilation issues, for another round of installation and tests.
