### 2021-01-08 #2946 @whedon
[REVIEW]: sboxgates: A program for finding low gate count implementations of S-boxes
**Submitting author:** @dansarie (<a href="http://orcid.org/0000-0001-9246-0263">Marcus Dansarie</a>)
**Repository:** <a href="https://github.com/dansarie/sboxgates" target ="_blank">https://github.com/dansarie/sboxgates</a>
**Version:** v2.0.0
