### 2021-01-08 #2946 @VivianePons
Hi, we now have two reviewers and we move the discussion here.

@gradvohl has already made some comments [over here, in the pre-review issue](https://github.com/openjournals/joss-reviews/issues/2932#issuecomment-756133402)  . @dansarie can you have a look?

Thank you very much to @gradvohl and  @gcdeshpande let me know if you have issues or questions
