### 2021-01-08 aef72e5
Lower required CMake version to 3.9.
