+++ CMakeLists.txt
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-cmake_minimum_required (VERSION 3.10)
+cmake_minimum_required (VERSION 3.9)
 
 project(SBOXGATES VERSION 1.0.0 LANGUAGES C)
 
