+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 # CMakeLists.txt
 #
-# Copyright (c) 2019-2020 Marcus Dansarie
+# Copyright (c) 2019-2021 Marcus Dansarie
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
