### 2021-01-08 0400754
programming language -> parallel computing platform
