### 2021-01-08 #2946 @dansarie
@VivianePons I looked over @gradvohl's comments and updated the repo. Responses below.

> * Tests: :-1:  (bad)

This did not come with any clarification. Testing is done by Travis CI automatically on each commit. I've updated the readme to make this clearer.

> ### Documentation in source code
>
> The author comments on some functions in the .h files, while other comments are in .c files. Also, most of the comments describe the functions at a high level only. It would be interesting to say what the parameters are and what they are needed for.
>
> **Suggestion**: Uniform where you will write the comments and, if possible, use a standard for documentation of functions, parameters, and returns (such as [Doxygen](https://www.doxygen.nl)).

All documentation of functions has been moved to header files and all parameters should now be documented.

> * When I tried to build the software with `cmake` it required the CMake 3.10 or higher.
>
> * **Suggestion**: Inform the user what are the required versions (especially CMake).

After testing, the required CMake version has been lowered to 3.9. The readme has been updated with information about the lowest required version.

> * When I tried to `cmake ..`, it raised the following error:
>   `-- Could NOT find MPI_C (missing: MPI_mpich_LIBRARY MPI_opa_LIBRARY MPI_mpl_LIBRARY MPI_C_WORKS)`
>
> * **Suggestion**: Inform the user what are the main environment variables must be set. Besides, I suspect that `mpich-devel` is also a dependency.

Installation of the `libmpich-dev` package is documented in the build section of the readme. CMake uses the script that is installed with the MPI implementation to search for MPI, so setting any environment variables shouldn't be required.

> ### Questions:
>
> When running with mpirun, may the user specify the number of processes and `hostfile` with processes location (for instance, `mpirun -np 4 --hostfile file ...`? If so, the author should clarify this in the instructions for running the software.

Yes. I have updated the readme with guidance on how to specify the number of processes and a reference to the MPI documentation.

>  1. Please consider to add the author's name and contact info in the `README.md`.

The readme has been updated with my name and a link to my Github profile.

>  2. In `paper.md`, the author states that CUDA is a "programming language for GPUs". However, CUDA is an Application Programming Interface (API) or a library, if you prefer.

I have changed this to "a parallel computing platform for Nvidia GPUs" in the paper.
