### 2021-01-09 #2946 @gradvohl
# Review itens for JOSS paper #2946
**Paper title**: sboxgates: A program for finding low gate count implementations of S-boxes
**Review round two**

## Review criteria
* A list of the authors of the software: :heavy_check_mark:
* Author affiliations: :heavy_check_mark:
* A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience: :heavy_check_mark:
* A clear statement of need that illustrates the purpose of the software: :heavy_check_mark:
* Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it: :heavy_check_mark:
* A list of key references including a link to the software archive:  :heavy_check_mark:

## Review items
* Software license: :thumbsup: (good - GPL v3)

* Documentation:
    * A statement of need: :thumbsup: (good)
    * Installation instructions: :thumbsup: (good)
    * Example usage: :thumbsup: (good)
    * API documentation: :thumbsup: (good)
    * Community guidelines: :thumbsup: (good)
    * Functionality: :thumbsup: (good)
    * Tests: :thumbsup: (good -- *It means that the software passed in my own tests, not the automatic ones*)

* Software paper
    * Authors: :heavy_check_mark:
    * A statement of need: :heavy_check_mark:
    * References: :heavy_check_mark:

## Reviewer comments
The author addressed all comments from Round 1. Therefore, I am satisfied with the software and with the documentation.
However, when I tried to run, it took much time to process the examples the author proposed. I know that the software addresses a complex computational problem. Still, I suggest that the author includes a section warning the users that the program may take high processing time to produce a result.
I ran the software on an AWS instance (t2.xlarge) with Ubuntu 20.04.

## Conclusion
Analyzing all the aspects mentioned earlier, I consider the paper **ready for publication**.
