### 2021-05-20 #2946 @HeshamAlsaadi
**I want to add the following comments:**
The references need to be built automatically from the .bib file (which will be added to the .pdf). Please follow this guideline here for implementation https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography; this will allow the citation to be very helpful and compiled correctly for a reference check.
