### 2021-05-23 #2946 @dansarie
@HeshamAlsaadi The references in the .bib file look ok to me and the reference check seems to be working.
