### 2021-06-15 #2946 @VivianePons
Hi @dansarie thank you for your patience, we can move toward acceptance!

Can you make a new tagged release of the software (if needed) an archive it (on Zenodo, figshare, or other)? Please be careful to use the same metadata than the paper (title and authors). Send me the doi when you are done!
