### 2021-06-15 #2946 @dansarie
@VivianePons done!

The tagged release (v2.0.0) is available at [https://github.com/dansarie/sboxgates/releases/tag/v2.0.0](https://github.com/dansarie/sboxgates/releases/tag/v2.0.0).
Version 2.0.0 has been archived at Zenodo with [doi:10.5281/zenodo.4954685](https://doi.org/10.5281/zenodo.4954685). (The latest version is always available through [doi:10.5281/zenodo.891020](https://doi.org/10.5281/zenodo.891020).)
