### 2021-06-16 #2946 @VivianePons
Thank you! We ow have everything to accept the paper into JOSS. Thanks a lot to @dansarie for the contribution and thank you @gradvohl and @HeshamAlsaadi for the reviews :)
